Refactor order property types in PdfProcess models and requests
- Changed 'orders.*' validation rule from 'numeric' to 'integer' in StorePdfProcessUploadRequest.
- Updated PdfProcessPage model to replace 'order' with 'tmp_order' as a float.
- Modified PdfProcessUpload model to change 'order' from float to integer.
- Adjusted upload method in PdfProcessUploadService to accept integer for order.
- Added migrations to update database schema for order properties in PdfProcessUpload and PdfProcessPage.

+++ app/Services/PdfProcessUploadService.php
@@ -7,7 +7,7 @@ use Illuminate\Http\UploadedFile;
 
 class PdfProcessUploadService
 {
-    public function upload(PdfProcess $pdfProcess, UploadedFile $file, float $order): void
+    public function upload(PdfProcess $pdfProcess, UploadedFile $file, int $order): void
     {
         $pdfProcess->uploads()->create([
             'name' => $file->getClientOriginalName(),
